requirements.txt
numpy>=1.26
pandas>=2.2
pyarrow>=15.0
PyYAML>=6.0
torch>=2.2
